# program/docs/api/openapi-college.json
{
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/v1/auth/login",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Unified API serving University, College, Secondary School, and Primary School management systems.",
    "title": "Education System API \u2014 College",
    "version": "1.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/attendance/course/{course_pk}/report": {
      "get": {
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Attendance \u2014 Course Report",
        "tags": [
          "Attendance"
        ]
      }
    },
...
    "/api/ilp": {
      "post": {
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Ilp \u2014 Create Plan",
        "tags": [
          "Ilp"
        ]
      }
    },
    "/api/ilp/{pk}": {
      "put": {
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Ilp \u2014 Update Plan",
        "tags": [
          "Ilp"
        ]
      }
    },
...
    "/api/messaging/{pk}/read": {
      "post": {
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Messaging \u2014 Mark Read",
        "tags": [
          "Messaging"
        ]
      }
    },
...
    "/api/notifications/{notification_id}/read": {
      "post": {
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Notifications \u2014 Mark Read",
        "tags": [
          "Notifications"
        ]
      }
    },
...
    "/api/reports": {
      "post": {
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Reports \u2014 Create Report",
        "tags": [
          "Reports"
        ]
      }
    },
    "/api/reports/{pk}": {
      "get": {
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Reports \u2014 Get Report",
        "tags": [
          "Reports"
        ]
      }
    },
...
    "/api/tlevel/routes": {
      "post": {
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Tlevel \u2014 Create Route",
        "tags": [
          "Tlevel"
        ]
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "servers": [
    {
      "description": "Current server",
      "url": "/"
    }
  ]
}